Another major cleanup of the castor code targetting to drop all warnings due to the -Wextra flag. On the way, many little bugs were fixed (possible core dumps in command lines, wrong or missing checks, ...), code was dropped since unused, files were converted from dos to unix end of line convention, etc, etc ...

+++ mediachanger/castorrmc/common/Cdomainname.c
@@ -47,7 +47,7 @@ int DLL_DECL Cdomainname(char *name, int namelen)
 					p++;
 				if (*p)
 					*(p + strlen (p) - 1) = '\0';
-				if (strlen (p) > namelen) {
+				if ((int)strlen (p) > namelen) {
 					serrno = EINVAL;
 					return (-1);
 				}
@@ -93,7 +93,7 @@ int DLL_DECL Cdomainname(char *name, int namelen)
 				if ((p = strchr (hp2->h_name, '.')) != NULL) {
 					free (haddrarray);
 					p++;
-					if (strlen (p) > namelen) {
+					if ((int)strlen (p) > namelen) {
 						serrno = EINVAL;
 						return (-1);
 					}
@@ -108,7 +108,7 @@ int DLL_DECL Cdomainname(char *name, int namelen)
 					if ((p = strchr (*hal, '.')) != NULL) {
 						free (haddrarray);
 						p++;
-						if (strlen (p) > namelen) {
+						if ((int)strlen (p) > namelen) {
 							serrno = EINVAL;
 							return (-1);
 						}
